fix: 마이페이지 게임 기록을 디벨롭 브랜치와 현재 브랜치 작업물 연동

+++ src/main/java/io/cavia/trader/module/member/repository/GameParticipationMapper.java
@@ -10,5 +10,5 @@ public interface GameParticipationMapper {
 
     List<GameParticipation> findByMemberId(Long memberId);
     List<GameParticipation> findByMemberIdWithPaging(Long memberId, int limit, int offset);
-    int countByMemberId(Long memberId);
+    Long countByMemberId(Long memberId);
 }
